+++ src/functions/objects-array-to-object.spec.ts
@@ -1,56 +1,5 @@
 import { objectsArrayToObject } from './objects-array-to-object';
 
-describe('objectsArrayToObject DEPRECATED', () => {
-
-    let consoleSpy: jest.SpyInstance;
-
-    beforeEach(() => {
-        consoleSpy = jest.spyOn(console, 'warn').mockImplementation();
-    });
-
-    afterEach(() => {
-        consoleSpy.mockRestore();
-    });
-
-    it('convert array to object', () => {
-        expect.assertions(3);
-        const array = [
-            {oneKey: 'oneValue', twoKey: 'twoValue'},
-            {oneKey: 'threeValue', twoKey: 'fourValue'},
-        ];
-
-        const result = objectsArrayToObject(array, 'oneKey');
-
-        expect(result).toStrictEqual({
-            oneValue: {oneKey: 'oneValue', twoKey: 'twoValue'},
-            threeValue: {oneKey: 'threeValue', twoKey: 'fourValue'},
-        });
-        expect(consoleSpy).toHaveBeenCalledTimes(1);
-        expect(consoleSpy).toHaveBeenCalledWith(
-            'Using objectsArrayToObject(T[], string) is deprecated, use a selector function as the second parameter.'
-        );
-    });
-
-    it('convert array to object with number keys', () => {
-        expect.assertions(3);
-        const array = [
-            {1: 'oneValue', 2: 'twoValue'},
-            {1: 'threeValue', 2: 'fourValue'},
-        ];
-
-        const result = objectsArrayToObject(array, '1');
-
-        expect(result).toStrictEqual({
-            oneValue: {1: 'oneValue', 2: 'twoValue'},
-            threeValue: {1: 'threeValue', 2: 'fourValue'},
-        });
-        expect(consoleSpy).toHaveBeenCalledTimes(1);
-        expect(consoleSpy).toHaveBeenCalledWith(
-            'Using objectsArrayToObject(T[], string) is deprecated, use a selector function as the second parameter.'
-        );
-    });
-});
-
 describe('objectsArrayToObject', () => {
     it('convert array to object', () => {
         expect.assertions(1);