+++ src/functions/objects-array-to-object.ts
@@ -1,18 +1,14 @@
-export const objectsArrayToObject = <T>(array: T[], attributeGetter: string | ((item: T) => string | number)) => {
-
-    let keyIsString = false;
-    if (typeof attributeGetter === 'string') {
-        // eslint-disable-next-line no-console
-        console.warn('Using objectsArrayToObject(T[], string) is deprecated, use a selector function as the second parameter.');
-        keyIsString = true;
-    }
+/**
+ * Convert an array containing objects to an object containing objects.
+ * @param array - The array to convert.
+ * @param attributeGetter - A function for getting the attribute to group on.
+ */
+export const objectsArrayToObject = <T>(array: T[], attributeGetter: (item: T) => string | number): { [index: string]: T } => {
 
     const object: { [index: string]: T } = {};
 
     for (const item of array) {
-        // @ts-ignore
-        const key = keyIsString ? item[attributeGetter] as string : attributeGetter(item);
-        object[key] = item;
+        object[attributeGetter(item)] = item;
     }
 
     return object;